// src/app/models/ReportModel.ts
export class BPCReportOV extends CommonClass {
    ID: number;
    Client: string;
    Company: string;
    Type: string;
    PatnerID: string;
    Material: string;
    GRR: string;
    MaterialText: string;
    InputQty: number;
    AccQty: number;
    RejQty: number;
    RejPercentage: number;
    Status: string;
}

export class PPMReportOption extends CommonClass {
    PartnerID: string;
    Material: string;
    PO: string;
    Status: string;
    FromDate: Date | string | null;
    ToDate: Date | string | null;
}

export class OverviewReportOption extends CommonClass {
    PartnerID: string;
    Material: string;
    PO: string;
    Status: string;
    FromDate: Date | string | null;
    ToDate: Date | string | null;
}